refactor: Fix paths
