+++ src/lib/Components/ArtworkFilter/Filters/ArtistIDsOptionsScreen.tsx
@@ -1,7 +1,7 @@
 import { StackScreenProps } from "@react-navigation/stack"
 import { ArtworksFiltersStore } from "lib/Components/ArtworkFilter/ArtworkFiltersStore"
 import React from "react"
-import { FilterModalNavigationStack } from "../ArtworkFilter"
+import { FilterModalNavigationStack } from "lib/Components/ArtworkFilter"
 import { ArtistIDsArtworksOptionsScreen } from "./ArtistIDsArtworksOptions"
 import { ArtistIDsSaleArtworksOptionsScreen } from "./ArtistIDsSaleArtworksOptionsScreen"
