+++ src/lib/Components/ArtworkFilter/Filters/ArtistIDsSaleArtworksOptionsScreen.tsx
@@ -7,7 +7,7 @@ import {
   FilterParamName,
 } from "lib/Components/ArtworkFilter/FilterArtworksHelpers"
 import React, { useState } from "react"
-import { FilterModalNavigationStack } from "../ArtworkFilter"
+import { FilterModalNavigationStack } from "lib/Components/ArtworkFilter"
 import { MultiSelectCheckOptionScreen } from "./MultiSelectCheckOption"
 
 interface ArtistIDsSaleArtworksOptionsScreenProps
