+++ src/lib/Components/ArtworkFilter/Filters/ColorsOptions.tsx
@@ -5,8 +5,8 @@ import { useLayout } from "lib/utils/useLayout"
 import { sortBy } from "lodash"
 import { Flex, space } from "palette"
 import React from "react"
-import { FilterModalNavigationStack } from "../ArtworkFilter"
-import { FancyModalHeader } from "../FancyModal/FancyModalHeader"
+import { FilterModalNavigationStack } from "lib/Components/ArtworkFilter"
+import { FancyModalHeader } from "lib/Components/FancyModal/FancyModalHeader"
 import { ColorsSwatch } from "./ColorsSwatch"
 import { useMultiSelect } from "./useMultiSelect"
