+++ src/lib/Components/ArtworkFilter/Filters/InstitutionOptions.tsx
@@ -9,7 +9,7 @@ import {
   ParamDefaultValues,
 } from "lib/Components/ArtworkFilter/FilterArtworksHelpers"
 import React from "react"
-import { FilterModalNavigationStack } from "../ArtworkFilter"
+import { FilterModalNavigationStack } from "lib/Components/ArtworkFilter"
 import { SingleSelectOptionScreen } from "./SingleSelectOption"
 
 interface InstitutionOptionsScreenProps
