+++ src/lib/Components/ArtworkFilter/Filters/MultiSelectOption.tsx
@@ -1,14 +1,14 @@
 import { ParamListBase } from "@react-navigation/native"
 import { StackNavigationProp } from "@react-navigation/stack"
 import { FilterData } from "lib/Components/ArtworkFilter/FilterArtworksHelpers"
-import { FilterToggleButton } from "lib/Components/ArtworkFilterOptions/FilterToggleButton"
+import { FilterToggleButton } from "lib/Components/ArtworkFilter/Filters/FilterToggleButton"
 import { useFeatureFlag } from "lib/store/GlobalStore"
 import { Box, Check, Flex, Sans, Separator, Text } from "palette"
 import React from "react"
 import { FlatList } from "react-native"
 import styled from "styled-components/native"
-import { FancyModalHeader } from "../FancyModal/FancyModalHeader"
-import { TouchableRow } from "../TouchableRow"
+import { FancyModalHeader } from "lib/Components/FancyModal/FancyModalHeader"
+import { TouchableRow } from "lib/Components/TouchableRow"
 
 interface MultiSelectOptionScreenProps {
   navigation: StackNavigationProp<ParamListBase>
