+++ src/lib/Components/ArtworkFilter/Filters/TimePeriodMultiOptions.tsx
@@ -9,7 +9,7 @@ import {
 import { FilterData } from "lib/Components/ArtworkFilter/FilterArtworksHelpers"
 import { useArtworkFiltersAggregation } from "lib/Components/ArtworkFilter/useArtworkFilters"
 import React, { useState } from "react"
-import { FilterModalNavigationStack } from "../ArtworkFilter"
+import { FilterModalNavigationStack } from "lib/Components/ArtworkFilter"
 import { MultiSelectOptionScreen } from "./MultiSelectOption"
 
 interface TimePeriodOptionsScreenProps
