+++ src/lib/Components/ArtworkFilter/Filters/__tests__/AdditionalGeneIDsOptions-tests.tsx
@@ -4,8 +4,8 @@ import { ArtworkFiltersState, ArtworkFiltersStoreProvider } from "lib/Components
 import { Aggregations, FilterParamName } from "lib/Components/ArtworkFilter/FilterArtworksHelpers"
 import React from "react"
 import { Switch } from "react-native"
-import { extractText } from "../../../../lib/tests/extractText"
-import { renderWithWrappers } from "../../../../lib/tests/renderWithWrappers"
+import { extractText } from "lib/tests/extractText"
+import { renderWithWrappers } from "lib/tests/renderWithWrappers"
 import { AdditionalGeneIDsOptionsScreen } from "../AdditionalGeneIDsOptions"
 import { OptionListItem } from "../MultiSelectOption"
 import { getEssentialProps } from "./helper"
