+++ src/lib/Components/ArtworkFilter/Filters/__tests__/AttributionClassOptions-tests.tsx
@@ -4,8 +4,8 @@ import { ArtworkFiltersState, ArtworkFiltersStoreProvider } from "lib/Components
 import { FilterParamName } from "lib/Components/ArtworkFilter/FilterArtworksHelpers"
 import React from "react"
 import { Switch } from "react-native"
-import { extractText } from "../../../../lib/tests/extractText"
-import { renderWithWrappers } from "../../../../lib/tests/renderWithWrappers"
+import { extractText } from "lib/tests/extractText"
+import { renderWithWrappers } from "lib/tests/renderWithWrappers"
 import { AttributionClassOptionsScreen } from "../AttributionClassOptions"
 import { OptionListItem } from "../MultiSelectOption"
 import { getEssentialProps } from "./helper"
