+++ src/lib/Components/ArtworkFilter/Filters/__tests__/TimePeriodMultiOptions-tests.tsx
@@ -6,8 +6,8 @@ import { FilterParamName } from "lib/Components/ArtworkFilter/FilterArtworksHelp
 import { __globalStoreTestUtils__ } from "lib/store/GlobalStore"
 import React from "react"
 import { Switch } from "react-native"
-import { extractText } from "../../../../lib/tests/extractText"
-import { renderWithWrappers } from "../../../../lib/tests/renderWithWrappers"
+import { extractText } from "lib/tests/extractText"
+import { renderWithWrappers } from "lib/tests/renderWithWrappers"
 import { OptionListItem as MultiSelectOptionListItem } from "../MultiSelectOption"
 import { TimePeriodMultiOptionsScreen } from "../TimePeriodMultiOptions"
 import { getEssentialProps } from "./helper"
